Refactor element finding logic to remove circular dependency
- Replace findElement function with findElementBySelectorAndText for improved clarity.
- Update error handling to provide more descriptive failure messages.
- Consolidate element finding logic into findStrategies.js for better organization.
